# 8281962: Avoid unnecessary native calls in InflaterInputStream