@simonis 2022-02-17
> Makes sense to me. Benchmark numbers look good.

Thanks Christoph!