@simonis 2022-02-17
> This change is probably okay 

Hi Alan,

thanks for looking at this change.

> but will require study to see if there are any side effects (sadly, this area has a history of side effects being reported months and years after a change). Would you mind holding off integrating this change until it has been reviewed by someone that works in the area?

Sure, no problem. But can you please be a little more specific on who you consider to qualify as "*works in the area*" :)