@LanceAndersen 2022-02-17
The change looks innocuous so it is probably OK.  I would like to kick of our Mach5 runs to see if it shakes out any potential issues.

From reading the 3rd party problem reports, it appears that the problem is with the 3rd party zlib implementations.  Hopefully this change will not mask other issues